change : cmds to cogs
