+++ cogs/bot/errors.py
