+++ cogs/bot/speedtest.py
