+++ cogs/entertainment/__init__.py
@@ -3,10 +3,10 @@ import traceback
 
 
 async def setup(bot):
-    for name in listdir("cmds/individual"):
+    for name in listdir("cogs/entertainment"):
         if not name.startswith(("_", ".")):
             try:
-                await bot.load_extension("cmds.individual."+name.replace(".py", ""))
+                await bot.load_extension("cogs.entertainment."+name.replace(".py", ""))
             except Exception as e:
                 print("[Log][err]" + str(e))
             else:
