+++ cogs/individual/__init__.py
@@ -1,0 +1,13 @@
+from os import listdir
+import traceback
+
+
+async def setup(bot):
+    for name in listdir("cogs/individual"):
+        if not name.startswith(("_", ".")):
+            try:
+                await bot.load_extension("cogs.individual."+name.replace(".py", ""))
+            except Exception as e:
+                print("[Log][err]" + str(e))
+            else:
+                print("[Log][load]" + name)
