+++ cogs/individual/obj_info.py
