+++ cogs/serverutil/__init__.py
@@ -3,10 +3,10 @@ import traceback
 
 
 async def setup(bot):
-    for name in listdir("cmds/entertainment"):
+    for name in listdir("cogs/serverutil"):
         if not name.startswith(("_", ".")):
             try:
-                await bot.load_extension("cmds.entertainment."+name.replace(".py", ""))
+                await bot.load_extension("cogs.serverutil."+name.replace(".py", ""))
             except Exception as e:
                 print("[Log][err]" + str(e))
             else:
